SolutionService/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
django-cors-headers==4.2.0
pika==1.3.2
uvicorn==0.30.1